feat: move components to pair device folder
