+++ src/pages/PairDevice/BeaconAddress.tsx
