+++ src/pages/PairDevice/ConnectExistingInctance.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react'
+import { withRouter } from 'storybook-addon-react-router-v6'
 
 import ConnectExistingInstance from './ConnectExistingInstance'
-import { withRouter } from 'storybook-addon-react-router-v6'
 
 const meta = {
   title: 'Connect-Device/ConnectExistingInstance',
